% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/c.eqsys.R
\name{c.eqsys}
\alias{c.eqsys}
\title{c.eqsys}
\usage{
\method{c}{eqsys}(...)
}
\arguments{
\item{...}{eqsys objects to be combined}
}
\value{
the new, combined eqsys object
}
\description{
Combine eqsys objects
}
\details{
Stacks up the equations and merges the lists of variables, parameters and shocks
}

