% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/derivVec.R
\name{derivVec}
\alias{derivVec}
\title{derivVec}
\usage{
derivVec(ex, x = attr(ex, "vlist"), param = vector("numeric", 0),
  xchk = function(z) {     TRUE })
}
\arguments{
\item{ex}{a vector of expressions, or an \code{eqsys} object}

\item{x}{a character vector of variable names w.r.t. which the expressions
will be differentiated.  If \code{ex} is an \code{eqsys} object,
the default is for \code{x} to be the \code{vlist} attribute of
\code{ex}.}

\item{param}{a vector of named parameter values that will be constant in
repeated calls to \code{fret} (the returned function)}

\item{xchk}{Range check function.  It returns \code{FALSE} if given an
               \code{x} vector that is outside the domain of definition of
               \code{ex}.
-------------------}
}
\value{
\code{fret}, a function that when evaluated at a numerical \code{x},
          returns the vector of expression values, but also, as
          \code{attr(value, "grad")}, the gradient matrix
          It uses the fixed \code{param} values set in the call to
          \code{derivVec}.
}
\description{
Application of \code{deriv()} to a vector
}

