% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/impulsgensys.R
\name{impulsgensys}
\alias{impulsgensys}
\title{impulsgensys}
\usage{
impulsgensys(gout, horiz, shocksize = NULL)
}
\arguments{
\item{gout}{a list of system matrices as produced by \code{gensys}}

\item{horiz}{number of periods over which to calculate responses}

\item{shocksize}{if non-NULL, a list of sizes of initial shocks}
}
\value{
a 3-d array, with dimensions variables x shocks x timespan, with
   the first two dimensions named.
}
\description{
Impulse responses for a solved linear RE model
}
\seealso{
\code{gensys}
}

