% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read.eqsys.R
\name{read.eqsys}
\alias{read.eqsys}
\title{read.eqsys}
\usage{
read.eqsys(file)
}
\arguments{
\item{file}{Usually a quoted file name, but for a small system
could be a text connection}
}
\value{
an \code{eqsys} object
}
\description{
Reads an equation system from a file or other connection
}

