% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sssys.R
\name{sssys}
\alias{sssys}
\title{sssys}
\usage{
sssys(eq)
}
\arguments{
\item{eq}{an eqsys object with a declared vlist attribute and with
lags denoted by l's attached to the end of the variable names.
(Notice that this means that if a variable name ends in l, it
should not be another variable name when the terminal l is removed.)}
}
\value{
A new system with all lagged variable names
  replaced by current variable names.
}
\description{
make steady state \code{eqsys} object from a dynamic one
}
\details{
Transform an \code{eqsys} dyanmic system to a static system for use
in finding the steady state, usually with \code{ssSolve()}.
}

